git: Refactor cgroups tests to use `gittest` package
The `git` package cannot use either the `testcfg` nor `gittest` packages
due to a cyclic dependency. As such, we have to do a lot of manual setup
in such tests.

Refactor the `cgroup` tests and move them into a separate `git_test`
package to fix this cyclic dependency. This simplifies the test setup
and allows us to easily create a test repository via the `gittest`
package so that we can switch from `gitCmdFactory.NewWithDir()` to
`New()`.