repository: Ignore license in subdirectory
The Go implementation was using git-ls-tree(1) with --full-tree to get
the whole tree of blobs at once. But this data isn't correctly passed on
to go-license-detector, so it never detected a license in a
subdirectory. Also, the Ruby implementation does not detect licenses in
subdirectories.

With this change we stop getting the whole tree of a repo, this stops
wasting resources on data that wasn't used.