Praefect: Remove clock drift readiness check
Currently the Praefect readiness RPC performs a clock drift check when
invoked. This clock drift check calls the configured NTP service to
validate the time against. If the NTP service call fails, the readiness
check is considered failed. The clock drift check should not be included
in the readiness check as it is more of a host problem. Also for
deployments that depend on the readiness RPC to signal readiness, a
failed response from the NTP service could inaccurately reflect the
state of Praefect.

This change removes the clock drift check from the checks performed by
the readiness RPC.