ruby: Remove version file
We have recently rewritten the way our Protobuf Gem is generated. As
part of that rewrite we have removed all generated Protobuf code from
our codebase as it is not needed anymore and instead generated ad-hoc
when building the Gem.

Part of this generated code is also `version.rb`, which contains the
current Gitaly version. This file is being written by the Release tools
bot when tagging a new release, but it is not really required as the
same version is also present in our `VERSION` file. We thus also started
to generate this file by seeding it with the contents of our `VERSION`
and have subsequently removed it from our tree in 326913168 (ruby:
Remove stale generated Protobuf code, 2022-10-27).

Given that the Release tools bot hasn't yet been upgraded though to stop
writing this file it was reintroduced in a later commit. The change to
the bot has since landed though, which effectively means that the file
is now stale and won't ever be updated anymore. So let's remove it.