Makefile: Simplify finding Go sources
In the past, there were Go code files other than the generated ones in
the `proto/` directory. That's not the case nowadays anymore as we have
since moved code for Protobuf-related tooling into `tools/`.

Simplify the condition to find our Go sources to just exclude all Go
files located in `proto/`. While at it, convert the excludes for the
top-level `_build` and `ruby` directory to also use `-path` so that we
don't accidentally skip any directories in Go-related subdirectories
that happen to be called like this. And finally, drop the exclude for
the `vendor` folder that we don't use anymore.