Makefile: Allow overriding the Protobuf code destination directory
Now that all generated Protobuf code lives in `proto/go` it is simple
enough for us to change the location of generated code to a different
destination. Introduce a variable that allows the caller to override
this location. This will be used by the subsequent commit, which changes
the logic for how we verify that generated code matches our definitions.