Parse invalid reference name errors in Updater
Our 'git update-ref' wrapper Updater doesn't currently parse the
the error messages and report invalid reference names with a typed
error. This commit adds support for that as it will be useful in
the future to report more useful errors when Gitaly verifies
references prior to accepting a transaction and logging it.