limithandler: Refactor tests to use `grpc_testing` gRPC service
The tests for our `limithandler` middleware use a custom gRPC service
definition to assert its behaviour. This is not necessary though as
`grpc-go` already provides a `grpc_testing` package that contains a gRPC
service that is designed for testing gRPC-specific behaviour.

Refactor the test to use this package and get rid of our own Protobuf
definitions.