Add a benchmark for TransactionManager's transaction throughput
This commit adds a benchmark that allows for measuring the transaction
processing throughput of the TransactionManager. The performance is
pretty abysmal for now given we're not optimizing for it yet. This does
give a baseline idea from where we are starting though and allows for
easily verifying the impact of future changes. There are a lot of low
hanging fruits such as using long-lived processes, processing the
transactions in batches and pipelining the various stages of execution.