housekeeping: Remove usage of `Time.AddDate(...)`
Currently we use `Time.AddDate(...)` in a few places in the housekeeping
package. The problem with this is that it directly adds/subtracts days
from the current time. This would not work well with DST changes. For
e.g. if you're subtracting day around the time of when DST changes
happen, you could potentially gain/loose an hour.

Instead lets use `Time.Add(...)` where we'd specify the number of hours
to be added/subtracted.