remote: Standardize errors creation
The error rate is tracked in the Sentry and if there are
context that is always different for the same error case
these errors won't be aggregated all together and won't
create a precedent to investigate. Those we change error
text to be same no matter what data is under processing.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/4471