Makefile: Fix race in `no-proto-changes` target
The `no-proto-changes` target first regenerates the Protobuf files and
then checks whether there are any changes compared to the Protobuf
definitions. To regenerate the files, we first delete the target
directory that contains the generated code so that we start from a clean
slate. This is causing problems though when this target is executed via
`make verify`, which besides other targets also executes the `lint`
target. Now if the code is getting deleted while the linter runs in
parallel then it may happen that linting fails due to the missing code.

Now that we have the ability to generate Protobuf code in a different
directory thoulgh this is easy to fix. Instead of modifying our source
directory, we can just generate Protobuf files in our build directory
and compare this directory with our sources. Like this, actual source
files files will never get deleted and thus the linter will be happy.