praefect: Remove global logger
Currently Praefect uses a global `logger` variable. When running
normally this is not an issue logged events occur serially.
However, in testing we try to run in parallel whenever possible, leading
to race-detector failures from events being logged concurrently. This
was occurring frequently in `TestAddRepositories_Exec()` in particular.

Make logger a local variable to resolve this.