operations: Structured errors for UserCommitFiles
The `UserCommitFiles` RPC may in some cases return a successful response
even in cases where an index operation or custom hook fails. This
creates unnecessary replication jobs for changes that did not go through
and causes our logs to misrepresent the request as successful.

To resolve this, return structured errors from this RPC.

Changelog: fixed