ci: Fix FIPS job not running for merge trains
The FIPS job is supposed to run whenever something gets merged onto the
default branch. This is supposed to happen by checking whether the
condition `CI_COMMIT_BRANCH == CI_DEFAULT_BRANCH` is true. But as it
turns out, this variable is not set in merge request pipelines.

Fix this by adding another condition that checks for the merge request
event type.