ci: Run macOS jobs when accepting merge requests
While macOS is not a supported platform for Gitaly, most of the
developers at GitLab do infact use macOS as their development platform.
As such it is sensible to not regress the ability of developers to
run the full test suite on macOS.

While it would be nicest to just always run the CI job, it does take
significantly longer to finish than our other tests. So instead, we do
the same as for our FIPS jobs and run them whenever a merge request gets
added to a merge train.