repository: Add benchmark for FindLicense
In the last commit we've swapped out spawning a git-cat-file(1) for the
use of ReadObject() from the localrepo package. At the moment the
implementation is nearly identical, but by using ReadObject() we'll
benefit from future improvements in that package.

To verify this change does not have negative performance implications,
this change adds a benchmark. I ran this benchmark several times on both
implementations, and there is no significant difference between the two.
Below is the output of one of the runs on my computer:

    goos: linux
    goarch: amd64
    pkg: gitlab.com/gitlab-org/gitaly/v15/internal/gitaly/service/repository
    cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
    BenchmarkFindLicense
    BenchmarkFindLicense/gitlab-org/gitlab.git
    BenchmarkFindLicense/gitlab-org/gitlab.git-8 30 37530506 ns/op
    BenchmarkFindLicense/stress.git
    BenchmarkFindLicense/stress.git-8             1 1104152375 ns/op