Add error to log.FieldsProducer interface
log.FieldsProducer is responsible for producing logging fields. gRPC
logging middleware trigger functions implementing this interface when
a RPC finishes. Before, they receive the context to generate fields to
add to the logs. In some cases, we need to produce the fields
differently between successful and failed RPCs. To support such cases,
we need to add error to the interface. The message producer is
responsible for passing RPC errors.