Attach feature_flags field to gRPC logs via FieldsProducer
In prior attempts, we tried to log feature_flags to Gitaly logs.
Unfortunately, we had to rollback immediately due to huge log volume.
This commit improves feature flag logging:
* Use FieldsProducer instead of gRPC middlewares. FieldsProducer is a
part of existing gRPC result logs. The logs are attached only once when
the RPC finishes.
* Log enabled flags only. Disabled flags can be implied indirectly
* Control logs via AlwaysLogFeatureFlag flag. By default, the field is
attached in RPC failures only. That field is attached to all logs if
this flag is on. Hopefully this will help controlling the log volume.