Return context canceled error if process is reaped due to cancellation
The command package sends a SIGTERM to a process if it's context is
canceled. This then results in the command terminating and returning
an error related to that. That error doesn't make it clear the reason
was the context cancellation. This commit changes Wait to return a
context canceled error if that was the reason of the process being
terminating.

The tests of the context cancellation are also updated to actually
wait until the process is running before cancelling the context.