repository: Allow controlling object format in CreateRepository
Allow callers to control the object format we pass to git-init(1) in the
`CreateRepository()` RPC. This feature is experimental and not supposed
to be used by any callers yet. We require this on the Gitaly side though
to increase our test coverage for callers of `gittest.CreateRepository`
that don't want to skip repository creation via gRPC services.

Changelog: added