helper: Pull out unused error code into a separate constant
Throughout the `helper.Err*` tests we need to verify behaviour when
wrapping an error that already has a gRPC error code assigned to it.
To avoid conflicts with the error code under test, we use an error code
that is not currently handled by any of our wrapping functions. This
code may start to get used as we add new wrappers though, and in that
case we have to go through all callsites to adapt the code.

Pull the unused error code out into a separate constant so that we can
easily change it.