gittest: Allow creating repos with non-default object format via gRPC
When `SkipCreationViaService` is not set in calls to `CreateRepository`,
we create the repository by calling the `CreateRepository()` RPC. As
this RPC didn't yet support creating repositories with a non-default
object format, this meant that this helper function didn't support not
skipping creation via the service when testing with the SHA256 object
format.

This limitation has now been fixed where the `CreateRepository` RPC now
has a new field `ObjectFormat` that allows the caller to override the
default object format. So let's enable creation via the service for
SHA256-based tests.