proto: Remove file_count from CommitLanguagesResponse
The CommitLanguagesResponse used to specify a file_count for each
Language, however the field was never filled in. It was originally
designed to be used, but due to performance issues it was rolled back in
de7a5d9d (Remove unused feature: File count commit languages,2020-07-06)

This change removes this field and adds a documentation string
indicating why the field is removed and providing a link to the issue.

Issue: https://gitlab.com/gitlab-org/gitaly/-/issues/4293