Verify correct usage of Updater
The updater can be in various states depending on what has been called
on it before. To retain the existing behavior, we want to ensure all
reference updates are done in transactions. If there is an error, it's
not clear whether or not the Updater is still usable afterwards. This
commit improves on this by checking the Updater is used in a correct
manner by erroring if the methods are called in unexpected order.
Additionally, the updater closes on encountering any sort of error.
This makes the calling semantics clear.