Allow non-unique addresses for storages in Praefect
Praefect has an artificial constraint where the configuration is
validated not to have multiple storages hosted by a given Gitaly.
This is not a real requirement. It should be possible to host
multiple storages in a virtual storage on a given Gitaly node. It's
not generally a good idea though given a single node doesn't offer
redundancy. There's shouldn't be a reason to fail hard on this, so
let's remove the check.

Changelog: fixed