Support multiple storages in Praefect testserver
Our helper for setting up a Praefect server in front of Gitaly nodes
in tests does not set up multiple virtual storages if the Gitaly nodes
have been configured with multiple storages. This makes the helper
unusable in tests which exercise multiple Gitaly storages which causes
us to not cover all of the behavior in tests. This commit configures a
virtual storage for each of the storages a test case configures for
Gitaly so we can better cover these cases with Praefect as well.