Enable Praefect in ReplicateRepository tests
ReplicateRepository tests are run without Praefect in front of the
Gitalys. This means we don't cover the RPC at all with Praefect which
has allowed a regression to slip through the tests. This commit enables
Praefect in these test cases and for now correctly asserts the behavior
divergence between Praefect and Gitaly. Due to relative path rewriting
in Praefect, the relative paths are equal validation check fails.

TestReplicateRepositoryTransactional still runs with Praefect disabled
as it is testing the transactional behavior and needs to override the
TransactionService. Same goes for TestFetchInternalRemote_successful.