Enable Praefect in TestRepositoryExists
TestRepositoryExists is ran with Praefect disabled. This allows for
behavior diverging between Praefect and Gitaly. This commit enables
Praefect in these tests and changes the assertions to accommodate for
the divergent behavior between Praefect and Gitaly.

Additionally, the repository is created without seeding it as the
content of the repository makes no difference for these tests.