Enable Praefect in TestRenameRepositorySuccess
TestRenameRepositorySuccess disables Praefect due to differing
behavior between Gitaly and Praefect. This commit enables Praefect
in the tests and changes the assertions to accommodate for the differing
behavior. This ensures the external API is tested through Praefect
as well.

The test was also explicitly deleting the new directory. That's not
necessary as our test helper clean up the temporary directory used
for the test.