Add Yamux parameters to Praefect's configuration
This commits makes some of the Yamux parameters configurable in order
to make it easier to find good values and later on support customizing
settings on a per installation level. For now, the parameters are only
parsed from the configuration but not yet applied. Later commit will
wire the parameters into the connections behind a feature flag.

Changelog: added