repository: Move Ruby sidecar tests into their own file
The Ruby sidecar tests are located in the `testhelper_test.go` file
right now. We're about to migrate some tests to SHA256, and that will
require us to remove the build constraint from that file. As the Ruby
sidecar test depends on calling sub-tests defined in other files that
would cause compilation to fail.

Move the test into a separate file to prepare for this refactoring.