git/stats: Accept `localrepo.Repo` when deriving object info
We're about to expose additional information in the git/stats package
that requires us to have access to the repository path. As we only get a
`git.RepositoryExecutor` as input right now that information is not
readily available though.

Convert the interfaces to accept a `localrepo.Repo` instead. We only
ever derive object info for local repositories, so there is no point in
using a more generic interface anyway.