housekeeping: Refactor code to derive config for writing commit-graphs
While the logic in `WriteCommitGraphForRepository()` works alright, it
really is quite confusing. First, we check whether the commit-graph
exists instead of checking for the commit-graph-chain. And second we
rely on the subtle behaviour of `IsMissingBloomFilters()` to return
`true` when there is a commit-graph, but it's not a chain.

Refactor the logic to check for existence of the commit-graph-chain. If
it doesn't exist or if it doesn't have bloom filters we'll rewrite the
complete commit-graph chain. This is a much more direct way of asking
for what really matters.