housekeeping: Explicitly consider existence of commit-graph chain
In housekeeping we will rewrite the commit-graph in case we don't have
bloom filters or in case there is no commit-graph-chain. While this
logic is correct, it is confusing that we don't explicitly handle the
case where we don't have a commit-graph chain at all.

Refactor the code to explicitly take into account existence of the
chain.