git/stats: Rewrite checking for commit-graph bloom filters
The `IsMissingBloomFilters()` function surprisingly doesn't check
whether we have bloom filters, but it verifies that we've got a
commit-graph _chain_ with bloom filters. So it will return `false` in
case we either have a monolithic commit-graph or no bloom filters.
This makes the function less usable in general and kind of surprising.

Refactor the function to instead return `CommitGraphInfo`, which is a
struct containing various information about the commit-graph. Besides
checking whether it contains bloom filters, it will also report whether
it exists in the first place and whether it is a commit-graph chain or
not. This gives callers a lot more information and it allows them to
stop manually checking for the existence of certain paths.