git: Remove linguist subcommand
In d6b9c7a9f (Add end-of-options to supported commands, 2020-05-19) we
added "linguist" as a subcommand for git that does not support
"--end-of-options". I don't fully understand why we ever added it,
because as far as I can tell we never called linguist as a git
subcommand.

Anyhow, in previous release we removed the use of the github-linguist
gem, so this subcommand is no longer relevant.