housekeeping: Start counting objects accurately
The housekeeping strategy is currently estimating the number of loose
objects by opening a single sharding directory, only, and then
extrapolating the number of objects in there. Now this extrapolation has
been a nice optimization. It is unlikely to really matter in practice
though as we only count objects in repository maintenance, which is
taking an extended amount of time anyway. And by becoming accurate, we
can eventually get rid of git-count-objects(1), where it will arguably
be more efficient to do the counting in-process than having to spawn a
separate process for this.

Count all shards to get an accurate object count. While at it, convert
the return value from an `int64` to an `uint64`.

Last but not least, add a benchmark to check how bad this is getting.
The following two results are for the old and new code, respectively:

    BenchmarkCountLooseObjects
    BenchmarkCountLooseObjects/empty_repository
    BenchmarkCountLooseObjects/empty_repository-16         	  106158	     11164 ns/op
    BenchmarkCountLooseObjects/repository_with_single_object
    BenchmarkCountLooseObjects/repository_with_single_object-16         	   80205	     14207 ns/op
    BenchmarkCountLooseObjects/repository_with_single_object_in_each_shard
    BenchmarkCountLooseObjects/repository_with_single_object_in_each_shard-16         	   75655	     14119 ns/op
    BenchmarkCountLooseObjects/repository_hitting_loose_object_limit
    BenchmarkCountLooseObjects/repository_hitting_loose_object_limit-16               	   64942	     19982 ns/op
    BenchmarkCountLooseObjects/repository_with_lots_of_objects
    BenchmarkCountLooseObjects/repository_with_lots_of_objects-16                     	     640	   1808591 ns/op

    BenchmarkCountLooseObjects
    BenchmarkCountLooseObjects/empty_repository
    BenchmarkCountLooseObjects/empty_repository-16         	    2720	    506955 ns/op
    BenchmarkCountLooseObjects/repository_with_single_object
    BenchmarkCountLooseObjects/repository_with_single_object-16         	    2371	    505337 ns/op
    BenchmarkCountLooseObjects/repository_with_single_object_in_each_shard
    BenchmarkCountLooseObjects/repository_with_single_object_in_each_shard-16         	     762	   1410647 ns/op
    BenchmarkCountLooseObjects/repository_hitting_loose_object_limit
    BenchmarkCountLooseObjects/repository_hitting_loose_object_limit-16               	     402	   2775585 ns/op
    BenchmarkCountLooseObjects/repository_with_lots_of_objects
    BenchmarkCountLooseObjects/repository_with_lots_of_objects-16                     	       2	 520969650 ns/op

So in the worst case we indeed look at a 100x slowdown. But we need to
be mindful of the context: this function is getting executed in the
context of repository housekeeping. Our P50 in this area is about 16ms,
which is the case where we decide that no housekeeping needs to take
place. With the new code we may now add 2.6ms to this P50 latency when
hitting the loose object limit, which is about 16% more latency. For P95
we have a flat bed of 75ms of latency, so with the 2.6ms we now have an
almost 4% hit to our latency. And for P99 latency we typically have
multiple seconds, where this overhead is really not going to matter at
all.

It is debatable whether this is acceptable or not. While we are slower
in the general case, it doesn't matter in the bigger picture where total
repository housekeeping time is dominated by the calls that take many
seconds to complete. And while not yet there, at some point this change
will allow us to drop a call to git-count-objects(1) in the housekeeping
code path, which should definitely make good for the performance hit we
have now.