git/stats: Report total object size when counting loose objects
As we iterate over the loose object sizes we can easily account for the
total size of loose objects as we already have that info available
anyway. Tracking this number here will eventually allow us to get rid of
the external git-count-objects(1) process, which currently does the book
keeping for us.

Convert the function to track the total loose object size. To avoid
confusion at callsites and to make this function more extensible, we
introduce a new structure `LooseObjectsInfo` that contains all relevant
information. To match existing naming in the package, this commit also
renames `CountLooseObjects()` to `LooseObjectsInfoForRepository()`.