housekeeping: Track total packfile size instead of largest one
The heuristical optimization strategy will repack objects less often
the bigger the largest packfile is. In hindsight, this is a rather weird
choice as repacking packfiles doesn't scale with the largest packfile,
but rather with the size of all packfiles combined.

Refactor the code to use the total packfile size as input instead of the
largest packfile size. While this will slightly change the frequency of
how often we do a full repack, I doubt it is overall going to matter
much.

This prepares for moving the packfile-counting logic into `git/stats`.