git/stats: Report garbage in loose object shards
Start reporting garbage files in loose object shards. This helps us to
abandon the external git-count-objects(1) process.

Note that this change fixes a bug with determining whether a loose
object name is valid or not. Previously, we used `LastIndexAny()` to
check whether the name only contains hex characters. This check fails
though in case the last character in the name is indeed a valid hex
character.