housekeeping: Fix treating object names as valid in some cases
We perform some very loose checks around whether object names are valid
or not when counting loose objetcs. To do so We Use `LastIndexAny()` to
check whether the name only contains hex characters. This check fails
though in case the last character in the name is indeed a valid hex
character.

Fix this bug by hand-rolling a new `isValidLooseObjectName()` function.
While using a regular expression would be neater, it would likely be a
whole lot slower than the hand-rolled loop.