Add --no-deref option to DeleteRefs RPC
By default, git-update-ref resolves the symbolic refs to their following
refs beforehand. As a result, deleting a symbolic ref results in
deleting its original ref. In most cases, this is not what we want. We
want to delete the symbolic ref itself instead. In addition, if both
symbolic ref and its following ref are in the same transaction, the
update fails to execute.

git-update-ref provides --no-deref tag. It disables ref chasing.
Symbolic refs themselves are modified. This is a perfect fit for
DeleteRef. However, it may be a bad idea to set it as the default value
for all ref updating operations.

Changelog: fixed
Issue: https://gitlab.com/gitlab-org/gitaly/-/issues/4624