git/stats: Use sub-structures to report loose and packfile info
Now that we have grown the ability to derive various bits of information
for loose objects and packfiles without spawning git-count-objects(1),
let's convert `ObjectsInfoForRepository()` to embed `LooseObjectsInfo`
and `PackfileInfo` structures instead of parsing them from the external
git-count-objects(1) process. This deduplicates logic and allows us to
get rid of spawning this process in the first place in a subsequent
commit.