structerr: Allow adding metadata to errors
Allow adding metadata to errors so that callers don't need to add
context to the error message, but can instead add it to the structured
error instead. The metadata that can be added is simply a map of keys
and values that allow us to add arbitrary information to any error
message.

In order to make this metadata useful, we need to get it out of the
system. This is done by amending the `statushandler.FieldsProducer`
function so that it will automatically amend any existing metadata as a
logged field. Consequentially, this metadata will now be part of the log
messages.

Ultimately this means that paradigms as following:

```go
return helper.ErrInvalidArgumentf("revision can't start with '-': %q", revision)
```

Would instead be written like this:

```go
return structerr.New("revision can't start with '-'").
    WithMetadata("revision", revision)
```