structerr: Introduce fields producer to log metadata
Introduce a new fields producer that allows us to inject error metadata
from the `structerr` package into logged errors and set up this producer
for the Gitaly server.

Note that we don't set this up for Praefect just yet. This is because
Praefect doesn't have the necessary logging infrastructure set up: it is
missing the `MessageProducer` as well as a `PerRPCLogHandler`, and these
are a prerequisite to install field producers. This is thus left for the
next iteration.