git/stats: Stop exposing commit-graph paths
Some of our tests in the repository service rely on the exposed
commit-graph paths in the `git/stats` package to check for the state of
commit-graphs in the repository. These can easily be converted to use
`CommitGraphInfoForRepository()` instead, which has the advantage that
we increase its test coverage and that the tests become more thorough.

Do so and stop exposing the commit-graph paths.