smarthttp: Improve error reporting in PostUploadPack
The error reporting in PostUploadPack is kind of lacking because we
don't use any breadcrumbs at all that would allow us to pin down where
an error has been raised. Combined with the fact that one of our tests
for this RPC doesn't verify returned error messages at all makes it
almost impossible to find out why that specific test is flaky in our CI
pipelines. We only see that something failed, but we cannot see right
now whether it fails in the expected way.

Improve the situation by adding breadcrumbs and checking whether the
returned error message matches our expectation. This hopefully helps us
to find out the root cause of flakiness, but if it doesn't it at least
improves our error handling.