smarthttp: Test for specific errors in PostUploadPack validation
The tests for validating requests to the PostUploadPack RPC only assert
error codes. This means that they may easily fail due to a completely
unrelated issue. And in fact, the last test that checks for requests
that contain non-empty data in the request header fails because the
repository wasn't found when testing with Praefect.

Improve the test to check the returned error message so that we can
fully nail down expected behaviour here. Furthermore, create a repo
against which we test the final test so that Praefect fails in the
expected way and not because of a missing repository database entry.