repository: Rewrite tests for OptimizeRepository
We have observed that tests for the OptimizeRepository RPC frequently
flake on macOS. Now I've been trying to find out why exactly that is
happening, but honestly the test is in such a bad shape that it's next
to impossible to figure out what's going on.

Rewrite the tests to exercise one thing at a time by using subtests.
This should make the tests more easily digestible and it should be
easier to figure out the state that can lead to failures given that we
now craft repositories tailored to the specific test.