housekeeping: Fix not rewriting commit-graphs when repacking
When repositories either don't have a commit-graph chain or when they
are lacking bloom filters then we need to rewrite the commit-graph so
that we create a new one that is both split and has bloom filters. This
logic doesn't kick in though in case a repository also needs its objects
to be repacked, as we would exit early in that case with a configuration
that doesn't ask for the commit-graph to be rewritten.

Fix this bug by reordering the conditions so that we first check for a
split commit-graph and bloom filters. Amend tests and add a new explicit
testcase for the optimization strategy. While at it, fix the description
of two tests to match their setup.