objectpool: Always reject fetching into nonexistent pools
With 2d2023a42 (objectpool: Reintroduce error if pool does not exist on
fetch, 2022-11-04), we have introduced a feature flag that causes us to
fail when FetchIntoObjectPool is executed with an object pool that does
not exist. This is done to fix multiple issues:

    - The lifetime of object pools is not clearly scoped, which makes it
      hard to reason about them.

    - Depending on whether an object pool has been created via
      CreatePool or via FetchIntoObjectPool, their state will be
      different.

    - Praefect would fail to properly handle object pools as the RPC is
      not marked as a repository-creating one.

We have rolled out these changes to production a week ago without any
negative consequences observed so far. So let's remove the feature flag
altogether to unconditionally fix all of the above issues.

Changelog: fixed